# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/user-account/congratulations": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "provider_name": {
                  "example": "any"
                },
                "user_id": {
                  "example": "any"
                },
                "user_name": {
                  "example": "any"
                },
                "user_email": {
                  "example": "any"
                },
                "user_phone": {
                  "example": "any"
                },
                "user_website": {
                  "example": "any"
                },
                "user_linkedin": {
                  "example": "any"
                },
                "user_github": {
                  "example": "any"
                },
                "user_picture": {
                  "example": "any"
                },
                "user_nickname": {
                  "example": "any"
                },
                "user_about": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user-account/edit-education": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "school_name": {
                  "example": "any"
                },
                "degree_name": {
                  "example": "any"
                },
                "start_date": {
                  "example": "any"
                },
                "end_date": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/user-account/edit-certificate": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "issuer_name": {
                  "example": "any"
                },
                "certificate_name": {
                  "example": "any"
                },
                "issue_date": {
                  "example": "any"
                },
                "expiration_date": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user-account/edit-experience": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "organization_name": {
                  "example": "any"
                },
                "position_name": {
                  "example": "any"
                },
                "start_date": {
                  "example": "any"
                },
                "end_date": {
                  "example": "any"
                },
                "employment_location_city": {
                  "example": "any"
                },
                "employment_location_state": {
                  "example": "any"
                },
                "employment_location_country": {
                  "example": "any"
                },
                "employment_type": {
                  "example": "any"
                },
                "employment_status": {
                  "example": "any"
                },
                "employment_description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user-account/edit-skill": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "skill_name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user-account/edit-language": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "language_name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user-account/edit-basic-info": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_about": {
                  "example": "any"
                },
                "user_email": {
                  "example": "any"
                },
                "user_phone": {
                  "example": "any"
                },
                "user_website": {
                  "example": "any"
                },
                "user_linkedin": {
                  "example": "any"
                },
                "user_github": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/user-account/update-account": {
      "get": {
        "description": "",
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}